fix(cache): resolve all golangci-lint errors
Fix 47 lint issues across cache and config packages:

- errcheck: Handle ignored error returns in olric.go
- gocritic: Pass large structs by pointer (Config, OlricConfig)
- gocritic: Fix import shadow (log -> olricLog)
- gocritic: Add named return values to parseBindAddr
- gocritic: Use bytes.Equal instead of string comparison
- govet: Reorder struct fields for better alignment
- godot: Add periods to comment endings
- misspell: Change "cancelled" to "canceled"
- revive: Mark unused test parameters with underscore
- unconvert: Remove unnecessary uint64 conversion
- errorlint: Use errors.Is for error comparison

+++ internal/cache/ristretto.go
@@ -263,7 +263,7 @@ func (r *ristrettoCache) Stats() Stats {
 		Hits:      metrics.Hits(),
 		Misses:    metrics.Misses(),
 		KeyCount:  metrics.KeysAdded() - metrics.KeysEvicted(),
-		BytesUsed: uint64(metrics.CostAdded() - metrics.CostEvicted()),
+		BytesUsed: metrics.CostAdded() - metrics.CostEvicted(),
 		Evictions: metrics.KeysEvicted(),
 	}
